feat: add current page functionality to the SideBar component (#5)
close #5
